requirements.txt
Flask
requests
pandas
matplotlib
SQLAlchemy
python-dotenv
pytest
